Выполнил задание по разбиению кода JS по модулям.
